requirements.txt
# Core
pandas==2.0.3
numpy==1.24.3
mysql-connector-python==8.1.0
sqlalchemy==2.0.23
pymysql==1.1.0

# Data Generation
Faker==19.3.0
tqdm==4.66.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Reporting
openpyxl==3.1.2
xlsxwriter==3.1.2
reportlab==4.0.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
colorama==0.4.6
tabulate==0.9.0